{{cookiecutter.project_slug}}/requirements.txt
Flask==2.3.2
werkzeug==2.3.4
jinja2==3.1.2
gunicorn==20.1.0
pytz==2023.3

psycopg==3.1.9
SQLAlchemy==2.0.13
SQLAlchemy-Utils==0.41.1
alembic==1.10.4

redis==4.5.5
celery==5.2.7

pytest==7.3.1
pytest-cov==4.0.0
flake8==6.0.0
isort==5.12.0
black==23.3.0
arrow==1.2.3
PyJWT==2.7.0
marshmallow-sqlalchemy==0.29.0
apifairy==1.3.0
Flask-DB==0.3.2
Flask-SQLAlchemy==3.0.3
Flask-Secrets==0.1.0
Flask-Static-Digest==0.4.0
flask-marshmallow==0.15.0
flask-httpauth==4.8.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-WTF==1.1.1
flask-debugtoolbar==0.13.1
email-validator==2.0.0
WTForms==3.0.1
